backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
python-dateutil>=2.8.0
celery>=5.3.0
redis>=4.6.0
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.2.0
